+++ create_schema.sql
@@ -200,6 +200,7 @@ CREATE TABLE dataset (
        annotation_source        VARCHAR(20),
        plot_default             VARCHAR(50), #options: 'bar', 'line', 'violin'
        annotation_release       INT,
+       assembly                 VARCHAR(20),
        FULLTEXT                 text_idx (title, ldesc),
        FULLTEXT                 text_with_geo_idx (title, ldesc, geo_id),
        FULLTEXT                 text_with_geo_pubmed_idx (title, ldesc, geo_id, pubmed_id),