+++ CHANGELOG.md
@@ -5,13 +5,15 @@ Changes to the DHART branch (gEAR) will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] - 2022-04
+## [Unreleased] - 2022-07
 
-### Added 
+### Added
+- Metadata assembly token.
 - h5ad uploader class.
 - Organisms: pig (img).
 
 ### Changed
+- Metadata annotation_source and annotation_release_number now optional.
 - CGI scripts (adjust shebang, see install notes).
 - MariaDB-related ini/setup files (create_schema.sql).
 - Organisms list (index.html).